main.py restructuration

+++ src/image.py
@@ -2,22 +2,25 @@
 
 """
 
-from os import listdir, path
-from os.path import isfile, join, getsize
+from os import listdir, path, mkdir
+from os.path import isfile, join, getsize, isdir
 from shutil import copy2
 
 
-def get_assets():
-    """Get windows assets and copy them to TEMP/ directory"""
+def get_assets(dst_dir='TEMP'):
+    """Get windows assets and copy them to a specific folder"""
     path_assets = path.expandvars(r'%LOCALAPPDATA%\Packages\Microsoft.Windows.ContentDeliveryManager_cw5n1h2txyewy\LocalState\Assets')
     tmp_assets = [f
                   for f in listdir(path_assets)
                   if isfile(join(path_assets, f))]
-    assets = []
+    new_assets = []
     for file in tmp_assets:
         if getsize(join(path_assets, file)) >= 10000:
-            assets.append(file)
+            # ? In order to get rid of logos that are in the same folder
+            new_assets.append(file)
             src = join(path_assets, file)
-            dst = 'TEMP/' + file + '.png'
-            copy2(src, dst)
-    return assets
+            dst = dst_dir + '/' + file + '.png'
+            if not isfile(dst):
+                # ? assets has not been collected yet
+                copy2(src, dst)
+    return new_assets
